requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
openai>=1.12.0
livekit
gunicorn==21.2.0

